% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst3k_function.R
\name{get_mst3k_quotes}
\alias{get_mst3k_quotes}
\title{Rowsdowser's Function (MST3K)}
\usage{
get_mst3k_quotes(id = sample(1:8106, 1))
}
\arguments{
\item{id}{Choose a numeric id. Defaults to a random id}
}
\description{
This function allows you to pull random snippets of Mystery Science Theater 3000 quotations
}
\examples{
get_mst3k_quotes()
}
\keyword{rowsdower}
